changed the startup window according to wireframe and few small chang…

+++ Sims2023/Sims2023/WPF/ViewModels/OwnerViewModel/AllRenovationsViewModel.cs
@@ -29,7 +29,8 @@ namespace Sims2023.WPF.ViewModels.OwnerViewModel
 
         public void Delete_Click()
         {
-            if (SelectedRenovation != null && SelectedRenovation.Status == "nije zapoceto")
+            TimeSpan difference = SelectedRenovation.StartDate - DateTime.Today;
+            if (SelectedRenovation != null && SelectedRenovation.Status == "nije zapoceto" && difference.TotalDays > 5)
             {
                 _accommodationRenovationService.Delete(SelectedRenovation);
                 renovations.Remove(SelectedRenovation);
